Revert default option to redirect output logs
Redirecting output logs for all BepInEx users has proven a bad practice, as it removes the log file from a place it is expected to be found by actual game developers. With the development of RedirectInternalLogs at BepInEx/BepInEx.Debug it is now possible to capture output logs without changing their location permanently, which satisfies the needs of both game developers and BepInEx users. As such, the default for output log redirection is better left reverted and instead the users should use the official RedirectInternalLogs preloader patcher as a safer alternative.